v2

+++ app/src/main/java/com/example/fbproject/MainActivity.kt
@@ -1,13 +1,18 @@
 package com.example.fbproject
 
 import android.R.attr.button
+import android.content.DialogInterface
 import android.content.Intent
 import android.os.Bundle
+import android.text.TextUtils
+import android.util.Log
 import android.view.View
 import android.widget.ImageButton
 import android.widget.PopupMenu
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.asLiveData
 import androidx.lifecycle.lifecycleScope
 import androidx.viewpager.widget.ViewPager
 import com.example.adapter.TabAdapter
@@ -21,47 +26,58 @@ class MainActivity : AppCompatActivity() {
 
     lateinit var userPreferences: UserPreferences
 
-    lateinit var nightMode: ImageButton
-    lateinit var dayMode: ImageButton
+    //lateinit var nightMode: ImageButton
+    //lateinit var dayMode: ImageButton
     override fun onCreate(savedInstanceState: Bundle?) {
         userPreferences = UserPreferences(this@MainActivity)
         super.onCreate(savedInstanceState)
+        val userPreferences = UserPreferences(this)
+        userPreferences.authToken.asLiveData().observe(this) {
+            Log.e("token################", it)
+            if (TextUtils.isEmpty(it) || it.equals("null") || it.isNullOrEmpty()) {
+                val intent = Intent(this, LoginActivity::class.java)
+                startActivity(intent)
+            }
+        }
         setContentView(R.layout.activity_main)
 
         menu.setOnClickListener {
             openOptionsMenu()
         }
 
-
-
-
-        menu.setOnClickListener { // Initializing the popup menu and giving the reference as current context
-            /*val popupMenu = PopupMenu(this@MainActivity, menu)
-
-            // Inflating popup menu from popup_menu.xml file
-            popupMenu.getMenuInflater().inflate(R.menu.main_menu, popupMenu.getMenu())
-            popupMenu.setOnMenuItemClickListener(object : MenuItem.OnMenuItemClickListener() {
-                override fun onMenuItemClick(menuItem: MenuItem): Boolean {
-                    // Toast message on menu item clicked
-                    Toast.makeText(this@MainActivity, "You Clicked " + menuItem.title, Toast.LENGTH_SHORT).show()
-                    return true
-                }
-            })
-            // Showing the popup menu
-            popupMenu.show()*/
-
+        menu.setOnClickListener {
             val popup = PopupMenu(this@MainActivity, menu)
-            popup.getMenuInflater().inflate(R.menu.main_menu, popup.getMenu());
+            popup.menuInflater.inflate(R.menu.main_menu, popup.menu);
             popup.setOnMenuItemClickListener(PopupMenu.OnMenuItemClickListener { item ->
                 when(item.itemId) {
-                    R.id.warrior ->
+                    R.id.warrior -> {
+                        val builder: AlertDialog.Builder = AlertDialog.Builder(this@MainActivity)
+                        builder.setMessage("Make me warrior")
+                        builder.setTitle("Alert !")
+                        builder.setCancelable(false)
+                        builder.setPositiveButton("I agree") { _: DialogInterface?, _: Int -> finish()}
+                        builder.setNegativeButton("Cancel") { dialog: DialogInterface, _: Int -> dialog.cancel() }
+
+                        val alertDialog: AlertDialog = builder.create()
+                        alertDialog.show()
                         Toast.makeText(this@MainActivity, "You Clicked : " + item.title, Toast.LENGTH_SHORT).show()
+                    }
                     R.id.logout ->{
-                        lifecycleScope.launch {
-                            userPreferences.deleteAuthToken()
+                        val builder: AlertDialog.Builder = AlertDialog.Builder(this@MainActivity)
+                        builder.setMessage("Do you want to Logout ?")
+                        builder.setTitle("Alert !")
+                        builder.setCancelable(false)
+                        builder.setPositiveButton("Yes") { _: DialogInterface?, _: Int -> finish()
+                            lifecycleScope.launch {
+                                userPreferences.deleteAuthToken()
+                            }
+                            val intent = Intent(this@MainActivity, LoginActivity::class.java)
+                            startActivity(intent)
                         }
-                        val intent = Intent(this@MainActivity, LoginActivity::class.java)
-                        startActivity(intent)
+                        builder.setNegativeButton("No") { dialog: DialogInterface, _: Int -> dialog.cancel() }
+
+                        val alertDialog: AlertDialog = builder.create()
+                        alertDialog.show()
                     }
 
                     R.id.edit_profile ->{
@@ -72,22 +88,14 @@ class MainActivity : AppCompatActivity() {
                 true
             })
             popup.show()
-
-
-
-
         }
 
 
-
-
-
-
-        var tabLayout = findViewById<TabLayout>(R.id.tabLayout)
-        nightMode = findViewById(R.id.night_mode)
-        dayMode = findViewById(R.id.day_mode)
-        var home = tabLayout.newTab()
-        var profile = tabLayout.newTab()
+        val tabLayout = findViewById<TabLayout>(R.id.tabLayout)
+        //nightMode = findViewById(R.id.night_mode)
+        //dayMode = findViewById(R.id.day_mode)
+        val home = tabLayout.newTab()
+        val profile = tabLayout.newTab()
         home.tag = "Home"
         profile.tag = "Profile"
         home.text = "Home"
@@ -98,8 +106,8 @@ class MainActivity : AppCompatActivity() {
 
         tabLayout.tabGravity = TabLayout.GRAVITY_FILL
 
-        var adapter : TabAdapter = TabAdapter(this,supportFragmentManager,tabLayout.tabCount)
-        var viewPager : ViewPager = findViewById(R.id.viewPager)
+        val adapter = TabAdapter(this,supportFragmentManager,tabLayout.tabCount)
+        val viewPager : ViewPager = findViewById(R.id.viewPager)
         viewPager.adapter = adapter
         viewPager.addOnPageChangeListener(TabLayout.TabLayoutOnPageChangeListener(tabLayout))
         tabLayout.addOnTabSelectedListener(object : TabLayout.OnTabSelectedListener {
@@ -115,7 +123,7 @@ class MainActivity : AppCompatActivity() {
             override fun onTabReselected(tab: TabLayout.Tab?) {
             }
         })
-        nightMode.setOnClickListener{
+        /*nightMode.setOnClickListener{
             //dayMode.isEnabled = true
             dayMode.visibility = View.VISIBLE
             nightMode.visibility = View.GONE
@@ -129,35 +137,18 @@ class MainActivity : AppCompatActivity() {
             dayMode.visibility = View.GONE
           //  dayMode.isEnabled = false
             //AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-        }
+        }*/
     }
-    /*override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        val inflater: MenuInflater = menuInflater
-        inflater.inflate(R.menu.main_menu, menu)
-        return true
-    }
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        userPreferences = UserPreferences(this@MainActivity)
-        // Handle item selection
-        return when (item.itemId) {
-            R.id.warrior -> {
+    override fun onBackPressed() {
+        val builder: AlertDialog.Builder = AlertDialog.Builder(this@MainActivity)
+        builder.setMessage("Do you want to exit ?")
+        builder.setTitle("Alert !")
+        builder.setCancelable(false)
+        builder.setPositiveButton("Exit") { _: DialogInterface?, _: Int -> finish()}
+        builder.setNegativeButton("Cancel") { dialog: DialogInterface, _: Int -> dialog.cancel() }
 
-                true
-            }
-            R.id.logout -> {
-                lifecycleScope.launch {
-                    userPreferences.deleteAuthToken()
-                }
-                val intent = Intent(this@MainActivity, LoginActivity::class.java)
-                startActivity(intent)
-                true
-            }
-            R.id.edit_profile -> {
-                val intent = Intent(this@MainActivity, EditProfileActivity::class.java)
-                startActivity(intent)
-                true
-            }
-            else -> super.onOptionsItemSelected(item)
-        }
-    }*/
+        val alertDialog: AlertDialog = builder.create()
+        alertDialog.show()
+    }
+
 }
